+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/domain/handlers/initiatetask/InitiateTaskEvaluateDmnResponse.java
@@ -1,13 +1,13 @@
-package uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.initiatetask;
+package uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.initiatetask;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common.DmnIntegerValue;
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common.DmnStringValue;
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.common.TaskEvaluateDmnResponse;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.DmnIntegerValue;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.DmnStringValue;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.common.TaskEvaluateDmnResponse;
 
 @ToString
 @EqualsAndHashCode(callSuper = true)
