+++ src/test/java/uk/gov/hmcts/reform/wacaseeventhandler/domain/initiatetask/InitiateTaskSendMessageRequestTest.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.wacaseeventhandler.domain.initiatetask;
 
 import org.junit.jupiter.api.Test;
 import pl.pojo.tester.api.assertion.Method;
-import uk.gov.hmcts.reform.wacaseeventhandler.domain.handler.initiatetask.InitiateTaskSendMessageRequest;
+import uk.gov.hmcts.reform.wacaseeventhandler.domain.handlers.initiatetask.InitiateTaskSendMessageRequest;
 
 import static pl.pojo.tester.api.assertion.Assertions.assertPojoMethodsFor;
